Rename in preparation for 'sessions'
