+++ server/routes/blockDatesOrSessions/blockDates/unblockDateController.test.ts
@@ -1,8 +1,8 @@
 import type { Express } from 'express'
 import request from 'supertest'
-import { appWithAllRoutes, flashProvider } from '../testutils/appSetup'
-import { createMockAuditService, createMockBlockedDatesService } from '../../services/testutils/mocks'
-import { MoJAlert } from '../../@types/bapv'
+import { appWithAllRoutes, flashProvider } from '../../testutils/appSetup'
+import { createMockAuditService, createMockBlockedDatesService } from '../../../services/testutils/mocks'
+import { MoJAlert } from '../../../@types/bapv'
 
 let app: Express
